% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\docType{data}
\name{animalList}
\alias{animalList}
\title{Some animal names}
\format{
A character vector with 277 items
}
\source{
<don't remember>
}
\usage{
animalList
}
\description{
Just a random list that is used w/ the toy dataset creator
}
\keyword{datasets}
