% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{functionsFinder}
\alias{functionsFinder}
\title{FUNCTION_TITLE}
\usage{
functionsFinder(paff = "R", matchingText = ".")
}
\arguments{
\item{paff}{PARAM_DESCRIPTION, Default: 'R'}

\item{matchingText}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
