% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{functionsUsedFinder}
\alias{functionsUsedFinder}
\title{find functions used in each of the functions in a specified R file}
\usage{
functionsUsedFinder(paff = "R", matchingText = ".")
}
\arguments{
\item{paff}{path to the file, Default: 'R'}

\item{matchingText}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function is suitable to identify what exports are required when
designing a package. Like a catch all in case sinew fails.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
