% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gitLister}
\alias{gitLister}
\title{figure out which folders have git initialized or not}
\usage{
gitLister(paff = ".")
}
\arguments{
\item{paff}{what folder contains all the folders that you want to confirm git-ness, Default: '.'}
}
\value{
returns a dataframe of all the folders that DO have git initialized.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1 \%>\%
 }
}
}
