% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gitOrganizer}
\alias{gitOrganizer}
\title{Identify which folders are git repos, runs `gc` on  git folders, and analyzes.}
\usage{
gitOrganizer(paff = "..", matchingText = ".")
}
\arguments{
\item{paff}{path to the file, Default: '..'}

\item{matchingText}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
will return a dataframe with the following fields: name,gitRepo (T/F),
branch(currently checked out branch), status (ok, behind or ahead of the remote),
toCommit (local changes).

Also, the console will print out folders that are NOT git repos, folders that ARE git
repos but are NOT on the master branch, and git repos that are in general not synced
with the remote.
}
\description{
This function will output a dataframe that will state whether folders
within a specified folder are git repos, and their status. By status I mean,
what's the currently checked-out branch, whether
there are any changes noticed between the local repo and the remote.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}}
}
