% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{igraphPlot}
\alias{igraphPlot}
\title{Create a prettier igraph chart}
\usage{
igraphPlot(
  g,
  size = "D",
  color = NULL,
  layout = "FR",
  nodeScale = 2,
  as = 0.1,
  keepBackbone = 0.5
)
}
\arguments{
\item{g}{graph from igraph}

\item{size}{What should govern node sizes ('D' for Degree, , 'B' for Betweenness,
'A' for Authority, 'H' for Hub, 'C' for Eigenvector Centrality), Default: 'D'}

\item{color}{What should govern the node colors? Same options as size (D,B,A,H),
Default: NULL}

\item{layout}{What layout should be used ('FR' for Fruchterman Reingold, 'G' for
Graphopt, 'K' for Kamada Kawai), Default: 'F'}

\item{nodeScale}{how much should we exaggerate the scale for node sizes, Default: 2}

\item{as}{Arrow head size, Default: 0.1}

\item{keepBackbone}{What percent of edges should be shown? We are using the
`layout_as_backbone` function from `graphlayouts` to trim less meaningful edges
for aesthetic purposes, Default = 0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1

 }
}
}
