% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{internalFunctionParser}
\alias{internalFunctionParser}
\title{this does most of the heavy lifting of seperating inputs and functions}
\usage{
internalFunctionParser(textString)
}
\arguments{
\item{textString}{what item should I break down}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
