% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{joiner}
\alias{joiner}
\title{internal function for testMatrixMaker}
\usage{
joiner(df)
}
\arguments{
\item{df}{dataframe representing each function to be parsed}
}
\value{
gives a df but with variables nicely labelled
}
\description{
starts operationalizing the equality of the params
}
\details{
this is probably not necessary, but in the off chance that
parameters are listed in roxygen in a different order than they are
listed in the function definition itself, it's safest to be super
explicit. This also makes mapping here and there easier.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
