% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{left}
\alias{left}
\title{gets left bit of string}
\usage{
left(text, num_char)
}
\arguments{
\item{text}{stuff to parse}

\item{num_char}{how many characters should be returned}
}
\value{
the `num_char` length of chars on the left of the inputed string
}
\description{
just like excel's left func
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 left("ABC", 2)
 }
}
}
