% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{libraryFinder}
\alias{libraryFinder}
\title{examines all the R files within a folder that have been modified within a certain date
and returns a list of all the libraries to install.}
\usage{
libraryFinder(parentFolder, howLong = 365, whatLib = .libPaths()[1])
}
\arguments{
\item{parentFolder}{what folder contains all the repos that contain the files to check}

\item{howLong}{specifies which files should be checked, by filtering to files only within
`howLong` days ago, Default: 365}

\item{whatLib}{what folder would you like to install these libraries into? It's wise
to install these in a new folder if you are upgrading to a new version of R, Default: .libPaths()[1]}
}
\value{
It spits out to console a command that you can copypaste to install all the CRAN libraries,
as well as listing all the non-CRAN libraries it found.
}
\description{
It'll check for all files that start with `.R`, so `.Rmd` etc are included.
}
\details{
For now the script is super naive, the next thing it can do is identify the package
dependencies and do a bit of a smarter install order.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
