% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{mapHelper}
\alias{mapHelper}
\title{A helper function to remember all the steps I like when mapping a
  throwaway function on somethin}
\usage{
mapHelper(objectName, functionName = "throwawayFunction1", parallel = FALSE)
}
\arguments{
\item{objectName}{what is the name of the list IN QUOTES that you're going to
be mapping? Eg "iris" not iris}

\item{functionName}{Optional, what is the name of the function you will be mapping?
Default: throwawayFunction1}

\item{parallel}{if your list contains many entries, set this to TRUE in order
to use multiple cores. If setting to TRUE gives an error, install the
package furrr}
}
\value{
returns nothing, but will output some text to the console suitable to
  be copied into your script
}
\description{
I always forget to set_names, to run safely, etc... so this is
  just to help save time. Assumes the tidyverse is loaded
}
\examples{
\dontrun{
if(interactive()){
 example <- list(a = 1, b = 2)
 mapHelper("example")
 }
}
}
