% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{mid}
\alias{mid}
\title{gets mid part of string}
\usage{
mid(text, start_num, num_char)
}
\arguments{
\item{text}{stuff to parse}

\item{start_num}{PARAM_DESCRIPTION}

\item{num_char}{how many characters should be returned}
}
\value{
the `num_char` length of chars in the middle of the inputed string
}
\description{
just like excel's mid func
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
