% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{packageBulkInstaller}
\alias{packageBulkInstaller}
\title{Build all packages within a folder that match a text string}
\usage{
packageBulkInstaller(paff, matchingText, onlyMaster = TRUE)
}
\arguments{
\item{paff}{path to the folder containing all the packages to build}

\item{matchingText}{text string that will be used to determine what packages to build}

\item{onlyMaster}{should we only build repos that are on MASTER branch, Default: TRUE}
}
\value{
returns an output of whether the installs were completed successfully or not.
}
\description{
When multiple people work on many repos, it it easy to see
if you are git pulled/pushed up, but it can be difficult or tedious to
see if you have built the latest version of each package. Therefore this
function helps install all repos that match a certain text-pattern.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
