% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{packageMakerList}
\alias{packageMakerList}
\title{Output a list of steps to do}
\usage{
packageMakerList()
}
\value{
just outputs to console stuff that should be done.
}
\description{
There are several steps to complete a package successfully. This function outputs a kind of
opinionated short-list or so
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
