% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{pleaseForTheLoveOfGodLetMeBuild}
\alias{pleaseForTheLoveOfGodLetMeBuild}
\title{remove intermediate build artifacts that make package development annoying}
\usage{
pleaseForTheLoveOfGodLetMeBuild(paff = "..")
}
\arguments{
\item{paff}{what path should we clean? It assumes you're inside a project, therefore
the default value, Default: '..'}
}
\value{
Nothing is returned, but the files/folders are deleted.
}
\description{
deletes the past builds by deleting every `tar.gz` file and `Rcheck`
folders in the specified paths. Also will go through every library folder and
delete all `LOCK` folders. Use this function when you are building a package but it
seems your new modifications aren't being implemented.
}
\details{
CARREFUL!! You're deleting stuff. Please check to make sure this function
does what you think it does.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
