% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{right}
\alias{right}
\title{gets right side of string}
\usage{
right(text, num_char)
}
\arguments{
\item{text}{stuff to parse}

\item{num_char}{how many characters should be returned}
}
\value{
the `num_char` length of chars on the right of the inputed string
}
\description{
just like excel's right func
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
