% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{testMatrixMaker}
\alias{testMatrixMaker}
\title{Create proposals for a testing matrix}
\usage{
testMatrixMaker()
}
\value{
will dump out a character vector of expressions suitable to
be paired with their testthat expectations.
}
\description{
Will go through all the R files in a package,
looking for instances of `testFrom` on roxygen2 `param` values in
function definitions. Once it finds these, it'll expand a grid of
all possibilities. This will provide a shortlist of all options to test
the code with, although it will not know if the input combination should
fail or pass, and if it does pass it should have a certain value.
}
\details{
This function is a bit brittle for now. It assumes that the
`testFrom` placement will be the last thing on param lines.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
