% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{text_collapser}
\alias{text_collapser}
\title{FUNCTION_TITLE}
\usage{
text_collapser(
  x,
  commas = TRUE,
  open_p = TRUE,
  close_p = TRUE,
  pipe = TRUE,
  no_double_lb = TRUE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{commas}{PARAM_DESCRIPTION, Default: TRUE}

\item{open_p}{PARAM_DESCRIPTION, Default: TRUE}

\item{close_p}{PARAM_DESCRIPTION, Default: TRUE}

\item{pipe}{PARAM_DESCRIPTION, Default: TRUE}

\item{no_double_lb}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
