% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{titleFlattener}
\alias{titleFlattener}
\title{Given a text with titles, it returns a df with each line with it's accompanying title}
\usage{
titleFlattener(text, titleChar)
}
\arguments{
\item{text}{the character string containing everything you're looking for.}

\item{titleChar}{what unique character denotes a title?}
}
\value{
returns a data frame with the erstwhile titles pushed into each row
}
\description{
This is suitable for when you have a long text and you're trying to do some NLP on it.
Some examples might be: trying to digest documentation, analyzing documentation, converting to a table
format something that should have never been entered into word, etc.
}
\details{
DETAILS
}
\examples{
 text <- c("## First","one","however","two","","## Second","otheer"," - Happy","Thoughts")
 titleFlattener(text, "##")
}
