% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{titleUnflattener}
\alias{titleUnflattener}
\title{given a df with two columns, stuff and section headers,
it outputs the stuff where with section headers are "titles"}
\usage{
titleUnflattener(textdf, spaceAbove = TRUE)
}
\arguments{
\item{textdf}{the df with the text and the section headers}

\item{spaceAbove}{should there be a line before the sections, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
basically this is the opposite of titleFlattener
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 text <- c("## First","one","however","two","","## Second","otheer"," - Happy","Thoughts")
 titleUnflattener(titleFlattener(text, "##"))
 }
}
}
