% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{track_dependencies}
\alias{track_dependencies}
\title{create a network chart of files in this folder}
\usage{
track_dependencies(remove_comments = TRUE)
}
\arguments{
\item{remove_comments}{by default, lines that have comments are thrown out,
but sometimes these can contain important information about objects that
have gotten created. In this case, the value should be explicitly stated as
FALSE #'}
}
\value{
plots a visnetwork object... doesn't return anything.
}
\description{
suitable to track file dependency, including data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
