% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{truncateNodes}
\alias{truncateNodes}
\title{filter a network according to simple node counts}
\usage{
truncateNodes(
  object,
  eliminateAbove = NULL,
  eliminateBelow = NULL,
  printResults = F
)
}
\arguments{
\item{object}{The list that contains nodes and edges, (the output of createNetwork)}

\item{eliminateAbove}{nodes larger than this number will be omitted, Default: NULL}

\item{eliminateBelow}{nodes smaller than this number will be omitted, Default: NULL}

\item{printResults}{View node degrees to help w/ trim limits?}
}
\value{
a list with 3 objects, same structure as the input
}
\description{
This is a dangerous thing to do if you're interested in statistics,
but for visualization purposes, it can be useful to make some decisions about what's
being removed. For the purposes of this function, the count is the node's degrees.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
