% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{internalFunctionParser}
\alias{internalFunctionParser}
\title{FUNCTION_TITLE}
\usage{
internalFunctionParser(textString)
}
\arguments{
\item{textString}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
