% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{mapHelper}
\alias{mapHelper}
\title{A helper function to remember all the steps I like when mapping a
  throwaway function on somethin}
\usage{
mapHelper(objectName)
}
\arguments{
\item{objectName}{what is the name of the list IN QUOTES that you're going to
be mapping? Eg "iris" not iris}
}
\value{
returns nothing, but will output some text to the console suitable to be copied into your script
}
\description{
I always forget to set_names, to run safely, etc... so this is
  just to help save time. Assumes the tidyverse is loaded
}
\examples{
\dontrun{
if(interactive()){
 goodstuff("iris")
 }
}
}
